refactor: move compose functions into their own module
